requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.21
psycopg2-binary==2.9.9
python-dotenv==0.21.0
pytest
pytest-cov
Flask-Migrate
Flask-Bcrypt==1.0.1
Werkzeug==3.0.4
bcrypt==4.2.0
blinker==1.8.2
click==8.1.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
importlib-metadata==8.5.0
zipp==3.20.2
applicationinsights==0.11.10
Flask-Login==0.6.3
